Restructuring Athena's module (#325)
* Splitting up athena's module in two files.

* General Athena's module refactoring.

* Updating tests.
