+++ awswrangler/athena/__init__.py
@@ -1,0 +1,13 @@
+"""Amazon Athena Module."""
+
+from awswrangler.athena._read import read_sql_query, read_sql_table  # noqa
+from awswrangler.athena._utils import (  # noqa
+    create_athena_bucket,
+    get_query_columns_types,
+    get_query_execution,
+    get_work_group,
+    repair_table,
+    start_query_execution,
+    stop_query_execution,
+    wait_query,
+)
