+++ docs/source/api.rst
@@ -80,6 +80,7 @@ Amazon Athena
 
     create_athena_bucket
     get_query_columns_types
+    get_query_execution
     get_work_group
     read_sql_query
     read_sql_table
