+++ tests/conftest.py
@@ -5,7 +5,7 @@ import pytest
 
 import awswrangler as wr
 
-from ._utils import extract_cloudformation_outputs, get_time_str_with_random_suffix, list_workgroups, path_generator
+from ._utils import create_workgroup, extract_cloudformation_outputs, get_time_str_with_random_suffix, path_generator
 
 
 @pytest.fixture(scope="session")
@@ -64,95 +64,67 @@ def loggroup(cloudformation_outputs):
 
 @pytest.fixture(scope="session")
 def workgroup0(bucket):
-    wkg_name = "aws_data_wrangler_0"
-    client = boto3.client("athena")
-    wkgs = list_workgroups()
-    wkgs = [x["Name"] for x in wkgs["WorkGroups"]]
-    if wkg_name not in wkgs:
-        client.create_work_group(
-            Name=wkg_name,
-            Configuration={
-                "ResultConfiguration": {"OutputLocation": f"s3://{bucket}/athena_workgroup0/"},
-                "EnforceWorkGroupConfiguration": True,
-                "PublishCloudWatchMetricsEnabled": True,
-                "BytesScannedCutoffPerQuery": 100_000_000,
-                "RequesterPaysEnabled": False,
-            },
-            Description="AWS Data Wrangler Test WorkGroup Number 0",
-        )
-    return wkg_name
+    return create_workgroup(
+        wkg_name="aws_data_wrangler_0",
+        config={
+            "ResultConfiguration": {"OutputLocation": f"s3://{bucket}/athena_workgroup0/"},
+            "EnforceWorkGroupConfiguration": True,
+            "PublishCloudWatchMetricsEnabled": True,
+            "BytesScannedCutoffPerQuery": 100_000_000,
+            "RequesterPaysEnabled": False,
+        },
+    )
 
 
 @pytest.fixture(scope="session")
 def workgroup1(bucket):
-    wkg_name = "aws_data_wrangler_1"
-    client = boto3.client("athena")
-    wkgs = list_workgroups()
-    wkgs = [x["Name"] for x in wkgs["WorkGroups"]]
-    if wkg_name not in wkgs:
-        client.create_work_group(
-            Name=wkg_name,
-            Configuration={
-                "ResultConfiguration": {
-                    "OutputLocation": f"s3://{bucket}/athena_workgroup1/",
-                    "EncryptionConfiguration": {"EncryptionOption": "SSE_S3"},
-                },
-                "EnforceWorkGroupConfiguration": True,
-                "PublishCloudWatchMetricsEnabled": True,
-                "BytesScannedCutoffPerQuery": 100_000_000,
-                "RequesterPaysEnabled": False,
+    return create_workgroup(
+        wkg_name="aws_data_wrangler_1",
+        config={
+            "ResultConfiguration": {
+                "OutputLocation": f"s3://{bucket}/athena_workgroup1/",
+                "EncryptionConfiguration": {"EncryptionOption": "SSE_S3"},
             },
-            Description="AWS Data Wrangler Test WorkGroup Number 1",
-        )
-    return wkg_name
+            "EnforceWorkGroupConfiguration": True,
+            "PublishCloudWatchMetricsEnabled": True,
+            "BytesScannedCutoffPerQuery": 100_000_000,
+            "RequesterPaysEnabled": False,
+        },
+    )
 
 
 @pytest.fixture(scope="session")
 def workgroup2(bucket, kms_key):
-    wkg_name = "aws_data_wrangler_2"
-    client = boto3.client("athena")
-    wkgs = list_workgroups()
-    wkgs = [x["Name"] for x in wkgs["WorkGroups"]]
-    if wkg_name not in wkgs:
-        client.create_work_group(
-            Name=wkg_name,
-            Configuration={
-                "ResultConfiguration": {
-                    "OutputLocation": f"s3://{bucket}/athena_workgroup2/",
-                    "EncryptionConfiguration": {"EncryptionOption": "SSE_KMS", "KmsKey": kms_key},
-                },
-                "EnforceWorkGroupConfiguration": False,
-                "PublishCloudWatchMetricsEnabled": True,
-                "BytesScannedCutoffPerQuery": 100_000_000,
-                "RequesterPaysEnabled": False,
+    return create_workgroup(
+        wkg_name="aws_data_wrangler_2",
+        config={
+            "ResultConfiguration": {
+                "OutputLocation": f"s3://{bucket}/athena_workgroup2/",
+                "EncryptionConfiguration": {"EncryptionOption": "SSE_KMS", "KmsKey": kms_key},
             },
-            Description="AWS Data Wrangler Test WorkGroup Number 2",
-        )
-    return wkg_name
+            "EnforceWorkGroupConfiguration": False,
+            "PublishCloudWatchMetricsEnabled": True,
+            "BytesScannedCutoffPerQuery": 100_000_000,
+            "RequesterPaysEnabled": False,
+        },
+    )
 
 
 @pytest.fixture(scope="session")
 def workgroup3(bucket, kms_key):
-    wkg_name = "aws_data_wrangler_3"
-    client = boto3.client("athena")
-    wkgs = list_workgroups()
-    wkgs = [x["Name"] for x in wkgs["WorkGroups"]]
-    if wkg_name not in wkgs:
-        client.create_work_group(
-            Name=wkg_name,
-            Configuration={
-                "ResultConfiguration": {
-                    "OutputLocation": f"s3://{bucket}/athena_workgroup3/",
-                    "EncryptionConfiguration": {"EncryptionOption": "SSE_KMS", "KmsKey": kms_key},
-                },
-                "EnforceWorkGroupConfiguration": True,
-                "PublishCloudWatchMetricsEnabled": True,
-                "BytesScannedCutoffPerQuery": 100_000_000,
-                "RequesterPaysEnabled": False,
+    return create_workgroup(
+        wkg_name="aws_data_wrangler_3",
+        config={
+            "ResultConfiguration": {
+                "OutputLocation": f"s3://{bucket}/athena_workgroup3/",
+                "EncryptionConfiguration": {"EncryptionOption": "SSE_KMS", "KmsKey": kms_key},
             },
-            Description="AWS Data Wrangler Test WorkGroup Number 3",
-        )
-    return wkg_name
+            "EnforceWorkGroupConfiguration": True,
+            "PublishCloudWatchMetricsEnabled": True,
+            "BytesScannedCutoffPerQuery": 100_000_000,
+            "RequesterPaysEnabled": False,
+        },
+    )
 
 
 @pytest.fixture(scope="session")
