+++ tests/test_athena_cache.py
@@ -51,14 +51,15 @@ def test_cache_query_ctas_approach_true(path, glue_database, glue_table):
     wr.s3.wait_objects_exist(paths=paths)
 
     with patch(
-        "awswrangler.athena._check_for_cached_results", return_value={"has_valid_cache": False}
+        "awswrangler.athena._read._check_for_cached_results",
+        return_value=wr.athena._read._CacheInfo(has_valid_cache=False),
     ) as mocked_cache_attempt:
         df2 = wr.athena.read_sql_table(glue_table, glue_database, ctas_approach=True, max_cache_seconds=0)
         mocked_cache_attempt.assert_called()
         assert df.shape == df2.shape
         assert df.c0.sum() == df2.c0.sum()
 
-    with patch("awswrangler.athena._resolve_query_without_cache") as resolve_no_cache:
+    with patch("awswrangler.athena._read._resolve_query_without_cache") as resolve_no_cache:
         df3 = wr.athena.read_sql_table(glue_table, glue_database, ctas_approach=True, max_cache_seconds=900)
         resolve_no_cache.assert_not_called()
         assert df.shape == df3.shape
@@ -81,14 +82,15 @@ def test_cache_query_ctas_approach_false(path, glue_database, glue_table):
     wr.s3.wait_objects_exist(paths=paths)
 
     with patch(
-        "awswrangler.athena._check_for_cached_results", return_value={"has_valid_cache": False}
+        "awswrangler.athena._read._check_for_cached_results",
+        return_value=wr.athena._read._CacheInfo(has_valid_cache=False),
     ) as mocked_cache_attempt:
         df2 = wr.athena.read_sql_table(glue_table, glue_database, ctas_approach=False, max_cache_seconds=0)
         mocked_cache_attempt.assert_called()
         assert df.shape == df2.shape
         assert df.c0.sum() == df2.c0.sum()
 
-    with patch("awswrangler.athena._resolve_query_without_cache") as resolve_no_cache:
+    with patch("awswrangler.athena._read._resolve_query_without_cache") as resolve_no_cache:
         df3 = wr.athena.read_sql_table(glue_table, glue_database, ctas_approach=False, max_cache_seconds=900)
         resolve_no_cache.assert_not_called()
         assert df.shape == df3.shape
@@ -103,7 +105,8 @@ def test_cache_query_semicolon(path, glue_database, glue_table):
     wr.s3.wait_objects_exist(paths=paths)
 
     with patch(
-        "awswrangler.athena._check_for_cached_results", return_value={"has_valid_cache": False}
+        "awswrangler.athena._read._check_for_cached_results",
+        return_value=wr.athena._read._CacheInfo(has_valid_cache=False),
     ) as mocked_cache_attempt:
         df2 = wr.athena.read_sql_query(
             f"SELECT * FROM {glue_table}", database=glue_database, ctas_approach=True, max_cache_seconds=0
@@ -112,7 +115,7 @@ def test_cache_query_semicolon(path, glue_database, glue_table):
         assert df.shape == df2.shape
         assert df.c0.sum() == df2.c0.sum()
 
-    with patch("awswrangler.athena._resolve_query_without_cache") as resolve_no_cache:
+    with patch("awswrangler.athena._read._resolve_query_without_cache") as resolve_no_cache:
         df3 = wr.athena.read_sql_query(
             f"SELECT * FROM {glue_table};", database=glue_database, ctas_approach=True, max_cache_seconds=900
         )
